+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [7.0.1] - 2020-05-17
 
 - Support results without a measurment name
+- use release influxdb client
 
 
 ## [7.0.0] - 2020-05-17